Removed *.rc file from setup, even on Windows.  Requires Windows SDK.
I'm trying to get this to build with only Visual Studio C++ Express, which does not come with
rc.exe.  The rc file was really to get the versioninfo block into the Windows DLL, but since
the extension is .pyd, it really isn't used anyway.  (Installers may use it, though?)  In the
near future, I'll add the version manually using ctypes.