Switch from SQLWCHAR to our own type.
Fixes #43.  unixODBC does not define the SQLWCHAR type to match its
requirements (2-byte chars), so I can't rely on it.  I've added
ODBCCHAR.

At this time it is hardcoded to 2-bytes which works for Windows and
unixODBC.  I need to add a mechanism to override this.

Tested on OS X postgres, Ubuntu 14 64-bix postgres, and Windows 7 64-bit
SQL Server.