SQLFreeHandle(stmt) can now raise an error if necessary.  Closes #74
In the past I ignored these since close is usually called during stack
unwinding and "real" errors shouldn't occur in free handle.  (That is,
the user of pyodbc can't cause errors here.)

Perhaps it is not true for this driver, however.  I still think it could
be something else, but it is better to raise errors when we can.  I do
ignore errors if we already have one, however.  I really don't want to
"eat" more important errors.