Rewrite Unicode handling for correctness and flexibility.
Reading and writing text now uses Python codecs which can be configured via the
new cnxn.setencoding and cnxn.setdecoding.  This is more complicated than simply
setting `cnxn.encoding = 'utf-16le'` but I've seen lots of weird combinations
that need to be handled more precisely.

This adds:

* connect(..., encoding=None)
* Connection.setencoding
* Connection.setdecoding

I'm not terribly happy with the API and would love some suggestions.  Perhaps we
should configure a "factory" with all of the settings that should be applied to
new connections?

I've also copied the docs from a separate branch into /docs which GitHub can
publish now.  I'm not sure how I want to handle documentation - there is a lot
of good information in the Wiki and I'm tempted to have minimal docs and cleanup
the Wiki.