Fix HSTMT leak.
For each new connection string (so usually once per app for most apps), pyodbc caches driver
configuration information.  During this an HSTMT was not freed.  Note that this should not have
had any affect since the call to SQLDisconnect later should free all handles.  (I may have even
not freed on purpose earlier, but I don't remember which is a good sign that closing it
explicitly would be better.)