Fix decoding of Cursor.description column names.
Fixes #190 - The issue was a column alias with an i with acute which caused an error.  pyodbc
now (1) uses the Unicode version of SQLDescribeColumn and (2) uses the configured SQL_C_WCHAR
encoding.  I've added SQL Server test decode_meta for this.

As part of this, I've moved much of the Unicode handling to textenc.h and textenc.cpp.  I fully
expect to replace SQLWChar in the future and move all handling to that file.