Rollback if exception raised in with clause.
The previous code would only commit at the end of a with clause if an exception was raised, but
it did not perform a rollback like the documentation stated.  While this is harmless if you are
allocating new connections in a scope so the automatic close would roll it back, it would be a
disaster if you were using a manual connection pool and didn't roll back yourself.