Changes for MS Access: use int instead of bigint; maxlength not set; add int tests
MS Access does not support BIGINT but the Python 3 build was using it exclusively.  It now uses
regular int when possible.

Since I modified the handling of Python longs, I added more tests to PostgreSQL and MySQL.

For some reason I could not reuse the HSTMT for multiple calls to SQLGetTypeInfo for
determining the maximum write size so I've changed it to allocate a new HSTMT for each.  This
makes it work by default for Access, but PostgreSQL reports a maximum write size of 255 which
is ridiculous and requires a manual override, documented in the Wiki.

Fixes #84