Fix some instances where Object wrapper copy ctor was being used.
I never intended for the Object copy ctor to be used and originally had it private.  I removed
it when a version of GCC failed to compile, but I've now put it back since it caused some
over decrementing.  I believe the GCC issue might have been because I didn't provide a body
for them (which used to work).

@dgsb provided a pull request for this months ago but I unwisely put it off because of the GCC
issue.  Unfortunately I put it off long enough that it didn't merge cleanly.

Closes #199 and #142