Clean up SQLite unit tests.
The tests had not been updated in a long while and were copied from another set of tests.  They
were using some older data types from pyodbc 2.x, for example.

Some of the tests were not even valid for SQLite.

Closes #95

There is still one failing test, but I'm concerned it is just the way SQLite works.  It seems
to start out in autocommit mode and the context manager fails.