Fix missing null terminators when encoding in UTF16 (#506) (#516)
This was noticed in Cursor.columns, but I'm sure it affects a couple of other places in the
code.

Unfortunately the solution is hideously inefficient.  I tried not null terminating, but I'm
afraid driver differences when not using SQLWCHAR, such as drivers that want UTF8, will cause
non-stop errors.  Suggestions for a more efficient implementation are welcome.  Alternatively,
if someone wants to try converting to using lengths instead of null terminated strings *and*
testing it, I'd be interested.