Fix UTF-16 conversion issues for connection errors on big endian
In #255, the code was changed to use the native endianness of the
system, but when a connection failure occurs the encoding was still
hard-coded to utf-16-le which causes mojibake on big endian systems.

Instead use ENCSTR_UTF16NE, which will be either utf-16-le or
utf-16-be, depending on the native endian for the system.

Fixes #563