Make Windows builds backwards compatible (#663)
Visual Studio 2019 included new C++ exception frame handling to reduce memory, but they put it
into a *new* DLL, so if you built with it you needed new DLLs installed.  Turn this off.  I
checked Python 3.8 and it is built without it also.  Eventually we'll have to turn the new
flags off when a version of Python is built with it.

Closes #663