Fix how tests find compiled libraries
The tests add the build directory to the path so we can test without installing.  Unfortunately
this is currently a manual search for the library as I'm not sure how to get it directly from
distutils.  In the past the search was limited to build directories for the current version,
such as "2.7" or "3.9".  This commit adds the system to it to differentiate between 32-bit and
64-bit builds on Windows.