Fix Python 2 encoding error (#941)
If a `str` object was passed to the SQLWChar constructor, the resulting pointer was zero.  The
zero was then passed to APIs like SQLColumns, resulting in either a crash or undefined
behavior.  (If the API accepts a zero, it wouldn't crash but often did something unexpected
like return multiple rows where one was expected.)

We now decode the bytes to Unicode then re-encode.