Remove duplicate type stub
`mypy` exits in an error state if the `pyodbc.pyi` file is included in a
type checking session, because of a duplicate definition of
`SQL_CONVERT_VARCHAR`.  This can be verified by installing `mypy` and
running:

    mypy src/pyodbc.pyi

This patch corrects the issue and allows the `pyodbc` usage to be
type-checked through at least use of these methods:

* `pyodbc.drivers()`
* `pyodbc.connect(...)`
* `pyodbc.connect(...).cursor()`
* `pyodbc.connect(...).cursor().__enter__()` (using a `with` context)
* `pyodbc.connect(...).cursor().__exit__()`
* `pyodbc.connect(...).cursor().execute()`
* `pyodbc.connect(...).cursor().fetchone()`

With a note for future work: `mypy --strict` against the same stub file
raises several errors pertaining to missing parameter types.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>