Rewrote Row.__repr__ due to memory errors
I ran the unit tests with a debug version of Python and it complained that the Unicode string I
was building was invalid.  The original code was a modified version of an older Python tuple
repr implementation, so I looked at doing that again.  However, cpython now uses an internal
_PyUnicode_Writer class we don't have access to, so I'm cheating by creating a tuple.

Since repr should not be in the critical path of most performance sensitive DB jobs, this will
do for now.