Rewrite Row.__repr__ to eliminate Unicode error
The debug build of Python generates an assertion that the resulting string we created has the
wrong type.  I think the issue was that we had a 1-byte string but it was not marked as ASCII
even though it had no values > 127.  Therefore I don't think it was causing problems or any
exploitable issues, but I'm not sure.

I replaced the implementation with simply building a temporary tuple with the values and
calling repr on them.  (Perhaps I should consider storing the rows in a Tuple instead of a
handcoded array?)

I also closed some popen pipes in the setup.py to eliminate some errors when PYTHONDEVMODE=1.