Rewrite decimal parsing to eliminate buffer overflow
This is a fix for GitHub security advisory GHSA-pm6v-h62r-rwx8.  The old code had a hardcoded
buffer of 100 bytes (and a comment asking why it was hardcoded!) and fetching a decimal greater
than 100 digits would cause a buffer overflow.

Author arturxedex128 supplied a very simple code to reproduce the error which was put into the
3 PostgreSQL unit tests as test_large_decimal.  (Thank you arturxedex128!)

Unfortunately the strategy is still that we have to parse decimals, but now Python strings /
Unicode objects are used so there is no arbitrary limit.