Comment out fastexecute temporarily.
I have not ported this code path and I'm not as familiar with it as I need to be.  To allow me
to complete porting and testing the rest, I've temporarily commented it out.

I will look into consolidating the binding for the two code paths.  Also, I'd like to consider
renaming it to "array binding" or "row wise binding" instead of "fast executemany".  While the
latter does tell us the goal of it, it is too generic.  For one thing, what if we wanted to
supply both row- and column-wise binding -- they are both "fast".