Rework output converter internals.  Delete unnecessary SQL Server test.
The unit test test_geometry_null_insert didn't really make sense so I've removed it.

I replaced the old output converter data structures with a Python dictionary.  I originally
used a custom array so I didn't have to construct a Python int object for the key.  Right now I
prefer the simplicity of using a built-in.  It will not affect performance of programs that
don't use output converters.  (We keep the dict* zero so we can quickly check it.)  Python
dictionaries are very optimized anyway.