Add support for running pyodbc unit tests without installing
For maximum dev turn around, I want to be able to run `setup build` and then `pytest ...`.  In
the past we had our tests scripts had their own main, so it was easy to do.  Now that we're
using pytest I had to add a conftest.py hook.

Since this could interfere with CI and some people just wont like it, it is only enabled if you
set PYODBC_TESTLOCAL=1 first.  I'm open to other ways of enabling it.