params: Move xdecrefs after last use
The code would work because the sequence still held references, but at first glance it looks
like it could be a bug.  Moving it makes it clear it is not.  An easier solution would be to
use the GET_ITEM macro which doesn't increment the reference count, but I'm hoping to move to
the stable ABI which doesn't support that.  We could also use the Object wrapper, but I want to
change as little as possible for now.