Build with and without NumPy support
This commit successfully builds `pyodbc` with and without NumPy support.
If NumPy is installed as an optional dependency, then you can retrieve
results from a database as NumPy arrays. If NumPy is not installed as an
optional dependency, then an error is given to the user indicating that
NumPy cannot be found.

- `setup.py` removes the `WITH_NUMPY` macro. `pyodbc` is set to build
  with NumPy support, as indicated in the `pyproject.toml` file. All
  references to this macro have also been removed in the C++ code.
- `src/cursor.cpp` removes the guards for allowing the `fetchdictarray`,
  on the cursor object, which is the way a user gets NumPy array objects
  back. It also aligns the objects in the `Cursor_methods` object.
- `src/npcontainer.cpp` removes NumPy build guards, moves the
  initialization of NumPy (with the import_array() macro).
- `src/npcontainer.h` removes unused guards.
- `src/pyodbcmodule.cpp` removes unused NumPy guards.